refactor: (#9) 인터페이스 분리
