+++ src/application/common/spi/axios.spi.ts
@@ -1,0 +1,6 @@
+import { ProjectGithubAxiosPort } from '../../domain/project/spi/project.spi';
+import { OAuthAxiosPort } from '../../domain/auth/spi/auth.spi';
+
+export interface AxiosPort extends ProjectGithubAxiosPort, OAuthAxiosPort {}
+
+export const AxiosPort = Symbol('IAxiosPort');
