--- src/application/github/spi/github.spi.ts
@@ -1,11 +1,0 @@
-import {
-  QueryRepositoryDetailsResponse,
-  QueryUserRepositoriesResponse
-} from '../../../infrastructure/thirdparty/github/dto/github.dto';
-
-export interface GithubPort {
-  getUserRepositories(username: string): Promise<QueryUserRepositoriesResponse[]>;
-  getRepositoryDetails(repositoryName: string): Promise<QueryRepositoryDetailsResponse>;
-}
-
-export const GithubPort = Symbol('IGithubPort');
