--- src/application/project/spi/project.spi.ts
@@ -1,7 +1,0 @@
-import { Project } from '../project';
-
-export interface ProjectPort {
-  saveProject(project: Project): Promise<Project>;
-}
-
-export const ProjectPort = Symbol('IProjectPort');
